Make issue package in icx package and move related modules
